Add community layer
Each route belongs to a community, the community source contains a
polygon feature per community which covers all of it's routes.

The layer is only intended for display at small scales so there is a
point style associated with the layer. The intention is that clicking on
one of these points should zoom the map to the extent of the community
which should be possible as the feature geometry is actually a polygon